+++ CHANGELOG.md
@@ -3,12 +3,13 @@ Changelog
 
 NOTE: isort follows the [semver](https://semver.org/) versioning standard.
 
-### 5.1.0 July 12, 2020
+### 5.1.0 July TBD, 2020
   - isort now throws an exception if an invalid settings path is given (issue #1174).
   - Fixed #1178: support for semicolons in decorators.
+  - Fixed #1315: Extra newline before comment with -n + --fss.
 
 ### 5.0.9 July 11, 2020
-  - Fixed #1301: Import headings in nested sections leads to check errors
+is
 
 ### 5.0.8 July 11, 2020
   - Fixed #1277 & #1278: New line detection issues on Windows.